requirements.txt
flask
openai
moviepy
requests
pillow
